+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] - 2026-04-07
+- Fixed: rrule fields default value
+
 ## [0.1.1] - 2026-03-17
 - Fixed: CSS class not applied to non-recurring events in reader
 - Fixed: backend asset loading